+++ client/src/main/java/eu/openanalytics/phaedra/curvedataservice/client/impl/HttpCurveDataServiceClient.java
@@ -72,7 +72,7 @@ public class HttpCurveDataServiceClient implements CurveDataServiceClient {
     }
 
     @Override
-    public CurveDTO createNewCurve(String substanceName, Long plateId, Long protocolId, Long featureId, Long resultSetId, double[] dose, double[] prediction) throws CurveUnresolvedException {
+    public CurveDTO createNewCurve(String substanceName, Long plateId, Long protocolId, Long featureId, Long resultSetId, float[] dose, float[] prediction) throws CurveUnresolvedException {
         var curveDTO = CurveDTO.builder()
                 .substanceName(substanceName)
                 .plateId(plateId)
