+++ dto/src/main/java/eu/openanalytics/curvedataservice/dto/CurveDTO.java
@@ -67,9 +67,9 @@ public class CurveDTO {
 
     private String version;
 
-    private double[] xAxisLabels;
+    private float[] xAxisLabels;
 
-    private double[] plotDoseData;
+    private float[] plotDoseData;
 
-    private double[] plotPredictionData;
+    private float[] plotPredictionData;
 }
